fix: temporarily disable actions language in CodeQL
CodeQL 2.23.3 appears to have an issue recognizing the 'actions' language
despite it being officially supported. Temporarily disabling it to allow
the Go, JavaScript/TypeScript, and Python analysis to proceed.

The 'actions' language can be re-enabled once the CodeQL issue is resolved
or when using a newer version that fixes this recognition problem.

Related: GitHub Actions workflows will still be analyzed when individual
workflow files are detected by other language analyzers.

Signed-off-by: Luca Muscariello <muscariello@ieee.org>